# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Garage API"
  description: "API for managing an auto garage, including customers, inventory, mechanics, and service tickets."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
definitions:
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: "Resource not found."
  ValidationErrorResponse:
    type: object
    properties:
      messages:
        type: object
        additionalProperties:
          type: array
          items:
            type: string
        example:
          email: ["Invalid email format.", "Email already exists."]
          password: ["Field may not be null."]
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: "Operation successful."

  # Customer Definitions
  CustomerCore:
    type: object
    properties:
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        format: email
        example: "john.doe@example.com"
      phone:
        type: string
        example: "555-1234"
    required:
      - name
      - email
  CustomerPayload:
    allOf:
      - $ref: "#/definitions/CustomerCore"
      - type: object
        properties:
          password:
            type: string
            format: password
            example: "securePassword123"
            writeOnly: true
        required:
          - password
  CustomerResponse:
    allOf:
      - $ref: "#/definitions/CustomerCore"
      - type: object
        properties:
          id:
            type: integer
            format: int64
            readOnly: true
            example: 1
          tickets:
            type: array
            items:
              $ref: "#/definitions/ServiceTicketSummary"
            readOnly: true
  CustomerLoginPayload:
    type: object
    properties:
      email:
        type: string
        format: email
        example: "john.doe@example.com"
      password:
        type: string
        format: password
        example: "securePassword123"
    required:
      - email
      - password
  CustomerLoginByIdPayload:
    type: object
    properties:
      password:
        type: string
        format: password
        example: "securePassword123"
    required:
      - password
  CustomerTicketsResponse:
    type: array
    items:
      $ref: "#/definitions/ServiceTicketResponse"

  # Mechanic Definitions
...
  InventoryItemSummary: # For nesting
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      name:
        type: string
        example: "Spark Plug"
      price:
        type: string
        example: "15.99"

  # Service Ticket Definitions
  ServiceTicketCore:
    type: object
    properties:
      vin:
        type: string
        example: "123ABC456DEF789GHI"
      service_date:
        type: string
        format: date
        example: "2024-07-26"
      service_desc:
        type: string
        example: "Oil change and tire rotation"
      customer_id:
        type: integer
        format: int64
        example: 1
    required:
      - vin
      - service_date
      - customer_id
...
  ServiceTicketResponse:
    allOf:
      - $ref: "#/definitions/ServiceTicketCore"
      - type: object
        properties:
          id:
            type: integer
            format: int64
            readOnly: true
            example: 1
          customer:
            $ref: "#/definitions/CustomerSummaryForTicket"
            readOnly: true
          mechanics:
            type: array
            items:
              $ref: "#/definitions/MechanicSummaryForTicket"
            readOnly: true
          inventory:
            type: array
            items:
              $ref: "#/definitions/ServiceTicketInventoryItemResponse"
            readOnly: true
  ServiceTicketSummary: # For nesting in Customer/Mechanic
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      vin:
        type: string
        example: "123ABC456DEF789GHI"
      service_date:
        type: string
        format: date
        example: "2024-07-26"
      service_desc:
        type: string
        example: "Oil change and tire rotation"
  CustomerSummaryForTicket: # For nesting in ServiceTicket
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        format: email
        example: "john.doe@example.com"
      phone:
        type: string
        example: "555-1234"
  MechanicSummaryForTicket: # For nesting in ServiceTicket
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      name:
        type: string
        example: "Jane Smith"
      email:
        type: string
        format: email
        example: "jane.smith@example.com"
      phone:
        type: string
        example: "555-5678"
      salary:
        type: string
        example: "60000.00"
  ServiceTicketInventoryItemBase:
    type: object
    properties:
      inventory_id:
        type: integer
        format: int64
        example: 1
      quantity:
        type: integer
        example: 2
    required:
      - inventory_id
      - quantity
  ServiceTicketInventoryItemResponse: # For ServiceTicket.inventory list
    allOf:
      - $ref: "#/definitions/ServiceTicketInventoryItemBase"
      - type: object
        properties:
          id: # This is ServiceInventory.id
            type: integer
            format: int64
            readOnly: true
            example: 101 
          inventory:
            $ref: "#/definitions/InventoryItemSummary"
            readOnly: true
  AuthTokenResponse:
    type: object
    properties:
      status:
        type: string
        example: "success"
      message:
        type: string
        example: "Successfully Logged In"
      auth_token:
        type: string
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login customer"
      description: "Authenticates a customer and returns an auth token."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerLoginPayload"
          examples:
            application/json:
              email: "john.doe@example.com"
              password: "securePassword123"
      responses:
        "200":
          description: "Login successful"
          schema:
            $ref: "#/definitions/AuthTokenResponse"
          examples:
            application/json:
              status: "success"
              message: "Successfully Logged In"
              auth_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        "400":
          description: "Invalid payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid payload, expecting email and password"
        "401":
          description: "Invalid email or password"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid email or password"
  /customers/login_id/{customer_id}:
    post:
      tags:
        - Customers
      summary: "Login customer by ID"
      description: "Authenticates a customer by ID and password, returns an auth token."
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          format: int64
          description: "ID of the customer to login"
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerLoginByIdPayload"
          examples:
            application/json:
              password: "securePassword123"
      responses:
        "200":
          description: "Login successful"
          schema: # The actual response includes a message and the AuthTokenResponse
            type: object
            properties:
              message:
                type: string
              response: # The route wraps the auth token response
                $ref: "#/definitions/AuthTokenResponse"
          examples:
            application/json:
              message: "Customer id: 1:"
              response:
                status: "success"
                message: "Successfully Logged In"
                auth_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        "400":
          description: "Invalid payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid payload, expecting password"
        "401":
          description: "Invalid ID or password"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid id or password"
  /customers/:
    post:
      tags:
        - Customers
      summary: "Create a new customer"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerPayload"
          examples:
            application/json:
              name: "Alice Wonderland"
              email: "alice.wonder@example.com"
              phone: "555-9876"
              password: "verySecurePassword"
      responses:
        "201":
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 2
              name: "Alice Wonderland"
              email: "alice.wonder@example.com"
              phone: "555-9876"
              tickets: []
        "400":
          description: "Validation error or email already exists"
          schema:
            oneOf: # Can be general error or validation error
              - $ref: "#/definitions/ErrorResponse"
              - $ref: "#/definitions/ValidationErrorResponse"
          examples:
            application/json: # Example for email exists
              error: "Email already associated with an account."
            application/json_validation: # Example for validation error
              messages:
                email: ["Not a valid email address."]
    get:
      tags:
        - Customers
      summary: "Get all customers"
      description: "Retrieves a list of all customers, with optional pagination."
      parameters:
        - name: page
          in: query
          type: integer
          description: "Page number for pagination"
        - name: per_page
          in: query
          type: integer
          description: "Number of items per page for pagination"
      responses:
        "200":
          description: "A list of customers"
          schema:
            type: array
            items:
              $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              - id: 1
                name: "John Doe"
                email: "john.doe@example.com"
                phone: "555-1234"
                tickets: []
              - id: 2
                name: "Alice Wonderland"
                email: "alice.wonder@example.com"
                phone: "555-9876"
                tickets: []
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Get customer by ID"
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: "Customer details"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john.doe@example.com"
              phone: "555-1234"
              tickets: [
                {
                  "id": 101,
                  "vin": "VIN123XYZ",
                  "service_date": "2024-07-28",
                  "service_desc": "Annual Checkup"
                }
              ]
        "404":
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Customer Not Found."
    put:
      tags:
        - Customers
      summary: "Update customer by ID"
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerPayload" # Password can be part of update
          examples:
            application/json:
              name: "Johnathan Doe"
              email: "john.doe.updated@example.com"
              phone: "555-1235"
              password: "newSecurePassword"
      responses:
        "200":
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              name: "Johnathan Doe"
              email: "john.doe.updated@example.com"
              phone: "555-1235"
              tickets: []
        "400":
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
          examples:
            application/json:
              messages:
                phone: ["Invalid phone number format."]
        "404":
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Customer not found."
    delete:
      tags:
        - Customers
      summary: "Delete customer by ID"
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: "Customer deleted successfully"
          schema:
            $ref: "#/definitions/MessageResponse"
          examples:
            application/json:
              message: "Customer id: 2 successfully deleted by user 1. Tickets reassigned to default customer (ID: 1)."
        "400":
          description: "Invalid customer ID format"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Invalid customer ID format in URL."
        "403":
          description: "Cannot delete the default customer"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Cannot delete the default customer account."
        "404":
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "Customer not found."
        "500":
          description: "System configuration error (default customer not found)"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              error: "System configuration error: Default customer not found. Please contact support."
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: "Get tickets for the authenticated customer"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "A list of service tickets for the customer"
          schema:
            $ref: "#/definitions/CustomerTicketsResponse"
          examples:
            application/json:
              - id: 101
                vin: "VIN123XYZ"
                service_date: "2024-07-28"
                service_desc: "Annual Checkup"
                customer_id: 1
                customer:
                  id: 1
                  name: "John Doe"
                  email: "john.doe@example.com"
                  phone: "555-1234"
                mechanics: [
                  {
                    id: 1,
                    name: "Mechanic Mike",
                    email: "mike@example.com",
                    phone: "555-0001",
                    salary: "70000.00"
                  }
                ]
                inventory: [
                  {
                    inventory_id: 1,
                    quantity: 1,
                    id: 201, # This is ServiceInventory.id
                    inventory: { # This is InventoryItemSummary
                      id: 1,
                      name: "Oil Filter",
                      price: "25.00"
                    }
                  }
                ]
        "404":
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json:
              message: "Customer not found" # Route returns 'message' not 'error'
...